requirements.txt
aiohttp==3.6.2
amqp==5.0.1
anyjson==0.3.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
billiard==3.6.3.0
cachetools==4.1.1
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.2
django-crispy-forms==1.9.2
django-filter==2.4.0
django-rest==0.8.6
djangorestframework==3.12.1
google-api-core==1.22.3
google-api-python-client==1.12.3
google-auth==1.22.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
idna-ssl==1.1.0
importlib-metadata==2.0.0
kombu==5.0.2
multidict==4.7.6
oauth2client==4.1.3
prompt-toolkit==3.0.7
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.14.0
pytz==2020.1
redis==3.4.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.0
zipp==3.2.0
